requirements.txt
flask>=3.0
werkzeug>=3.0
pywebview>=5.0
